requirements.txt
Flask==1.0.2
firebase_admin
PyYAML
feedparser
google-cloud-tasks==1.1.0
googleapis-common-protos==1.6.0
ffmpeg-python
youtube-dl
gcs-oauth2-boto-plugin
gsutil
